Some more clean up.
