+++ addons/spectator/functions/fn_onLoad.sqf
@@ -3,10 +3,7 @@ params ["_display"];
 #include "\x\tmf\addons\spectator\script_component.hpp"
 
 
-
-
 with uiNamespace do {
-
     GVAR(display) = _display;
     GVAR(unitlabel) = _display displayCtrl IDC_SPECTATOR_TMF_SPECTATOR_UNITLABEL;
     GVAR(unitlist) = _display displayCtrl IDC_SPECTATOR_TMF_SPECTATOR_UNITLIST;
@@ -19,28 +16,25 @@ with uiNamespace do {
     GVAR(view) = _display displayCtrl IDC_SPECTATOR_TMF_SPECTATOR_VIEW;
     GVAR(mute) = _display displayCtrl IDC_SPECTATOR_TMF_SPECTATOR_MUTE;
     GVAR(map) = _display displayCtrl IDC_SPECTATOR_TMF_SPECTATOR_MAP;
-
-    GVAR(compass) = _display displayCtrl IDC_SPECTATOR_TMF_SPECTATOR_COMPASS;
-    GVAR(compassL) = _display displayCtrl IDC_SPECTATOR_TMF_SPECTATOR_COMPASSLEFT;
-    GVAR(compassR) = _display displayCtrl IDC_SPECTATOR_TMF_SPECTATOR_COMPASSRight;
+    GVAR(compass) = [_display displayCtrl IDC_SPECTATOR_TMF_SPECTATOR_COMPASS,_display displayCtrl IDC_SPECTATOR_TMF_SPECTATOR_COMPASSLEFT,_display displayCtrl IDC_SPECTATOR_TMF_SPECTATOR_COMPASSRight];
+    GVAR(compass) = 
+    GVAR(compassL) = ;
+    GVAR(compassR) = ;
 
     _labelParent = _display displayCtrl 2300;
     GVAR(labels) = [
-    _labelParent controlsGroupCtrl 6
-    ,_labelParent controlsGroupCtrl 7
-    ,_labelParent controlsGroupCtrl 8
-    ,_labelParent controlsGroupCtrl 9
-    ,_labelParent controlsGroupCtrl 10
-    ,_labelParent controlsGroupCtrl 11
+        _labelParent controlsGroupCtrl 6,
+        _labelParent controlsGroupCtrl 7,
+        _labelParent controlsGroupCtrl 8,
+        _labelParent controlsGroupCtrl 9,
+        _labelParent controlsGroupCtrl 10,
+        _labelParent controlsGroupCtrl 11
     ];
-
 
     GVAR(map) ctrlShow (missionNamespace getVariable [QGVAR(showMap),false]);
     GVAR(map) mapCenterOnCamera false;
 
 };
-
-
 
 
 if(!getMissionConfigValue ["TMF_Spectator_AllSides",true]) then {
@@ -59,7 +53,6 @@ _allowedModes = [getMissionConfigValue ["TMF_Spectator_AllowFollowCam",true],get
 
 // if ACRE2 is enabled, enable the mute button
 if (isClass(configFile >> "CfgPatches" >> "acre_main")) then {
-
     [true] call acre_api_fnc_setSpectator;
     _data = ["ACRE2","HeadSet"] call CBA_fnc_getKeybind;
     GVAR(mute_key) = (_data select 5) select 0;
@@ -73,10 +66,8 @@ if (isClass(configFile >> "CfgPatches" >> "acre_main")) then {
         } forEach tmf_acre2_languagesTable;
         _languages call acre_api_fnc_babelSetSpokenLanguages;
     };
-
 }
-else // else remove it
-{
+else { // else remove it
     with uiNamespace do {
         GVAR(mute) ctrlShow false; // hide mute button
     };
